more py3 fixin
