+++ tests/functional/test_proxy.py
@@ -4,15 +4,15 @@ import sys
 import pexpect
 
 from vncdotool import rfb
-
+from .helpers import PExpectAssertMixin
 
 class TestLogEvents(TestCase):
     def setUp(self):
-        cmd = 'vncev -rfbport 5999 -rfbwait 1000'
+        cmd = u'vncev -rfbport 5999 -rfbwait 1000'
         self.server = pexpect.spawn(cmd, timeout=2)
         self.server.logfile_read = sys.stdout
 
-        cmd = 'vnclog --listen 1842 -s :99 -'
+        cmd = u'vnclog --listen 1842 -s :99 -'
         self.recorder = pexpect.spawn(cmd, timeout=2)
         self.recorder.logfile_read = sys.stdout
 
@@ -22,23 +22,11 @@ class TestLogEvents(TestCase):
             self.recorder.terminate(force=True)
 
     def run_vncdo(self, commands):
-        cmd = 'vncdo -s localhost::1842 ' + commands
+        cmd = u'vncdo -s localhost::1842 ' + commands
         vnc = pexpect.spawn(cmd, timeout=2)
         vnc.logfile_read = sys.stdout
         retval = vnc.wait()
         assert retval == 0, (retval, str(vnc))
-
-    def assertKeyDown(self, key):
-        down = '^.*down:\s+\(%s\)\r' % hex(key)
-        self.server.expect(down)
-
-    def assertKeyUp(self, key):
-        up = '^.*up:\s+\(%s\)\r'  % hex(key)
-        self.server.expect(up)
-
-    def assertMouse(self, x, y, buttonmask):
-        output = '^.*Ptr: mouse button mask %s at %d,%d' % (hex(buttonmask), x, y)
-        self.server.expect(output)
 
     def test_key_alpha(self):
         self.run_vncdo('key z')
